Create README markdown and reorder sections
Began compiling results and discussion in a README.md file to replace the README.tst file. Put the modeling for win times immediately after the modeling for average times, followed by a summary table of performance for all models. Put the coefficient analysis for factors affecting average times and win times in a single table at the end.